+++ test/movers_test.go
@@ -1,11 +1,13 @@
-package movers
+package test
 
 import (
 	"testing"
+
+	"github.com/samjtro/go-tda/movers"
 )
 
 func TestMovers(t *testing.T) {
-	_, err := Get("$DJI", "up", "percent")
+	_, err := movers.Get("$DJI", "up", "percent")
 
 	if err != nil {
 		t.Fatalf(err.Error())